Fixing the namespaces and moving all the parsers (task #5394)
To simplify code management all the parsers are grouped
under same Type, not under different objects as it was before.

It'll decrease the code duplication drastically.
