+++ src/ObjectType/Attendee.php
@@ -1,0 +1,9 @@
+<?php
+namespace Qobo\Calendar\ObjectType;
+
+class Attendee
+{
+    protected $id;
+
+    protected $name;
+}
