+++ src/ObjectType/Parsers/Cake/AttendeeParser.php
@@ -1,0 +1,13 @@
+<?php
+namespace Qobo\Calendar\ObjectType\Parsers;
+
+use Qobo\Calendar\ObjectType\Attendee;
+use Qobo\Calendar\ObjectType\Parsers\ParserInterface;
+
+class AttendeeParser implements ParserInterface
+{
+    public function parse($data)
+    {
+
+    }
+}
